requirements.txt
google-generativeai
faiss-cpu
sentence-transformers
PyPDF2
python-dotenv
numpy
langchain
langchain-community
tiktoken
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
python-multipart
